Change vagrant box and options
The Ubuntu cloud image doesn't use the default virtual box guest
additions. So if the VirtualBox host is running a new version updating
this is difficult.

This commit also changes the VM shared folder permissions which cause
issues with Ansible.

Additionally, we're increasing the memory of the logging server since
elasticsearch and logstash don't like small amounts of memory.