Update grafana web interface configuration
Use the influxdb options to save interfaces into a influxdb database.

This role and the influxdb will need to change the usage of the
username/password/database variables as currently these are integrated.

Ideally the influxdb role will create the databases and users using the
URI module and have the grafana roles share and use these values.